Extract shared PyICeJSONEncoder to lab_utils.
Consolidates the duplicate JSON encoders from plugin_manager
(_CrashLogJSONEncoder) and test_results (CustomJSONizer) into a
single PyICeJSONEncoder in lab_utils/json_encoder.py. Handles numpy
arrays/scalars, datetimes, timedeltas, and bytes with a repr()
fallback that never raises.
